+++ src/main/kotlin/cloud/glitchdev/rfu/feature/mob/BossHealthBar.kt
@@ -1,7 +1,7 @@
 package cloud.glitchdev.rfu.feature.mob
 
-import cloud.glitchdev.rfu.config.categories.GeneralFishing
-import cloud.glitchdev.rfu.config.categories.GeneralFishing.HEALTH_BAR_REGEX
+import cloud.glitchdev.rfu.config.categories.RareScSettings
+import cloud.glitchdev.rfu.config.categories.RareScSettings.HEALTH_BAR_REGEX
 import cloud.glitchdev.rfu.events.managers.MobEvents.registerMobDetectEvent
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
@@ -12,13 +12,13 @@ import cloud.glitchdev.rfu.manager.mob.MobManager
 object BossHealthBar : Feature {
     override fun onInitialize() {
         registerMobDetectEvent { entities ->
-            if(!GeneralFishing.bossHealthBars) {
+            if(!RareScSettings.bossHealthBars) {
                 BossHealthBarDisplay.updateEntities(emptySet())
                 return@registerMobDetectEvent
             }
             val filteredEntities = entities.filter { HEALTH_BAR_REGEX.matches(it.sbName) }
             BossHealthBarDisplay.updateEntities(filteredEntities.toSet())
-            if(GeneralFishing.boostPollingRate) MobManager.boostDetectionRate(filteredEntities.isNotEmpty())
+            if(RareScSettings.boostPollingRate) MobManager.boostDetectionRate(filteredEntities.isNotEmpty())
         }
     }
 }