+++ src/main/kotlin/cloud/glitchdev/rfu/feature/mob/TimeToKill.kt
@@ -1,7 +1,7 @@
 package cloud.glitchdev.rfu.feature.mob
 
-import cloud.glitchdev.rfu.config.categories.GeneralFishing
-import cloud.glitchdev.rfu.config.categories.GeneralFishing.RARE_SC_REGEX
+import cloud.glitchdev.rfu.config.categories.RareScSettings
+import cloud.glitchdev.rfu.config.categories.RareScSettings.RARE_SC_REGEX
 import cloud.glitchdev.rfu.constants.text.TextColor
 import cloud.glitchdev.rfu.events.managers.MobEvents.registerMobDisposeEvent
 import cloud.glitchdev.rfu.feature.Feature
@@ -15,7 +15,7 @@ import kotlin.time.Clock
 object TimeToKill : Feature {
     override fun onInitialize() {
         registerMobDisposeEvent { entities ->
-            if(!GeneralFishing.timeToKill) return@registerMobDisposeEvent
+            if(!RareScSettings.timeToKill) return@registerMobDisposeEvent
             val entities = entities.filter { RARE_SC_REGEX.matches(it.sbName) }
             entities.forEach { entity ->
                 val duration = Clock.System.now() - entity.createdAt